EllipticCurves/ProSign-3/requirements.txt
ecdsa==0.19.0
pycryptodome==3.20
